Add DbContexct, BaseRepositoty,Logic in (Country,City,Package,Addition) Controllers

+++ BoatRentalSystem.API/Controllers/AdditionController.cs
@@ -1,4 +1,6 @@
-﻿using BoatRentalSystem.Application;
+﻿using AutoMapper;
+using BoatRentalSystem.API.ViewModel;
+using BoatRentalSystem.Application.Services;
 using BoatRentalSystem.Core.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -10,43 +12,73 @@ namespace BoatRentalSystem.API.Controllers
     public class AdditionController : ControllerBase
     {
         private readonly AdditionService _additionService;
+        private readonly IMapper _mapper;
 
-        public AdditionController(AdditionService additionService)
+        public AdditionController(AdditionService additionService, IMapper mapper)
         {
             _additionService = additionService;
+            _mapper = mapper;
         }
 
         [HttpGet]
-        public Task<IEnumerable<Addition>> Get()
+        public async Task<ActionResult<IEnumerable<AdditionViewModel>>> Get()
         {
-            return _additionService.GetAllAdditions();
+            var city = await _additionService.GetAllAdditions();
+            var cityViewModel = _mapper.Map<IEnumerable<AdditionViewModel>>(city);
+
+            return Ok(cityViewModel);
         }
-           
 
         [HttpGet("{id}")]
-
-        public Task<Addition> Get(int id)
+        public async Task<ActionResult<AdditionViewModel>> GetById(int id)
         {
-            return _additionService.GetAdditionById(id);
+            var city = await _additionService.GetAdditionById(id);
+            if (city == null)
+            {
+                return NotFound();
+            }
+            var cityViewModel = _mapper.Map<AdditionViewModel>(city);
+
+            return Ok(cityViewModel);
         }
 
         [HttpPost]
-        public Task Post(Addition addition)
+        public async Task<ActionResult> Post([FromBody] AddAdditionViewModel addAdditionViewModel)
         {
-            return _additionService.AddAddition(addition);  
+            var Addition = _mapper.Map<Addition>(addAdditionViewModel);
+            await _additionService.AddAddition(Addition);
+            return CreatedAtAction(nameof(Get), new { id = Addition.Id }, addAdditionViewModel);
         }
 
         [HttpPut]
-            public Task Put(Addition addition)
+        public async Task<ActionResult> Put(AdditionViewModel AdditionViewModel)
         {
-                return _additionService.UpdateAddition(addition);
+            var existingAddition = await _additionService.GetAdditionById(AdditionViewModel.Id);
+            if (existingAddition == null)
+            {
+                return NotFound();
+            }
+            // Not Working (Why???????????????????????????)
+            //var city = _mapper.Map<City>(cityViewModel);
+            //await _additionService.UpdateCity(city);
+            //return Ok(city);
+            _mapper.Map(AdditionViewModel, existingAddition);
+            await _additionService.UpdateAddition(existingAddition);
+            return Ok(existingAddition);
 
         }
         [HttpDelete]
-
-        public Task Delete(int id)
+        public async Task<ActionResult> Delete(int id)
         {
-            return _additionService.DeleteAddition(id);
+            var existingCity = await _additionService.GetAdditionById(id);
+            if (existingCity == null)
+            {
+                return NotFound();
             }
+            await _additionService.DeleteAddition(id);
+            return NoContent();
+        }
+
+
     }
 }
